Criando novas entidades

+++ src/main/java/com/bradesco/minhasfinancas/servicos/UsuarioService.java
@@ -12,5 +12,5 @@ public interface UsuarioService {
 
     void validarEmail(String email);
 
-    Optional<Usuario> obterPorId(Long id);
+    Optional<Usuario> obterPorId(Integer id);
 }
